Added validation for order date
Admins were selecting order dates in the future to have their properties showing first in the index page. That Washington a problem because if you add a property today, and there are lots of properties with order dates in the future, then the property added today will show after them. The main idea of order_date id to bring back to the first page properties that were published in the past. Therefore, the use that the admins were giving to this field Washington incorrect. By adding the order_date_cannot_be_in_the_future validation, I intend to restrain admins from using this field incorrectly.