Made property price smaller
The price was exceeding the div’s width on tablet.
I added styles to fix that.
I also added a sass mixins stylesheet to add a media queries mixing and
save some time when writing media queries.
Some prices were wrapped inside a .col-xs-6 and others weren’t. So I
standardized prices to be wrapped in a .col-xs-x always