Changes to avoid deleting public description
When Vitaly did the changes to enable multi-language, a problem
emerged. Whenever a user created a property, when they hit save, the
public description field was not saved.
I removed the code added in the controller related to the description
selected depending on the locale.
I also removed repeated fields from rails_admin property panel.