Fix problem with property codigo sequence
Some properties were created for one admin and assigned to other admin.
That created an issue with the admin codigo sequence, because when a new
property was created, the app would look for all admin properties and
order them by code. But since there were properties with other code, it
was creating conflicts.
The solution was to filter only properties that started with the admin
code.